+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/AddFromTemplateCommand.cs
@@ -1,21 +1,18 @@
 ﻿namespace Sitecore.FakeDb.Data.Engines.DataCommands
 {
-  using Sitecore.Data.Items;
-  using Sitecore.Diagnostics;
   using System.Threading;
+  using Sitecore.Data.Items;
 
   public class AddFromTemplateCommand : Sitecore.Data.Engines.DataCommands.AddFromTemplateCommand, IDataEngineCommand
   {
-    private ThreadLocal<DataEngineCommand> innerCommand;
+    private readonly ThreadLocal<DataEngineCommand> innerCommand;
 
     private ThreadLocal<ItemCreator> itemCreator;
 
     public AddFromTemplateCommand()
     {
+      this.innerCommand = new ThreadLocal<DataEngineCommand> { Value = DataEngineCommand.NotInitialized };
       this.itemCreator = new ThreadLocal<ItemCreator>();
-
-      this.innerCommand = new ThreadLocal<DataEngineCommand>();
-      this.innerCommand.Value = DataEngineCommand.NotInitialized;
     }
 
     public virtual void Initialize(DataEngineCommand command)