+++ Source/Sitecore.FakeDb/Data/Engines/DataCommands/ResolvePathCommand.cs
@@ -2,18 +2,17 @@
 {
   using System;
   using System.Linq;
+  using System.Threading;
   using Sitecore.Data;
   using Sitecore.Diagnostics;
-  using System.Threading;
 
   public class ResolvePathCommand : Sitecore.Data.Engines.DataCommands.ResolvePathCommand, IDataEngineCommand
   {
-    private ThreadLocal<DataEngineCommand> innerCommand;
+    private readonly ThreadLocal<DataEngineCommand> innerCommand;
 
     public ResolvePathCommand()
     {
-      this.innerCommand = new ThreadLocal<DataEngineCommand>();
-      this.innerCommand.Value = DataEngineCommand.NotInitialized;
+      this.innerCommand = new ThreadLocal<DataEngineCommand> { Value = DataEngineCommand.NotInitialized };
     }
 
     public virtual void Initialize(DataEngineCommand command)